minor review feedback
