+++ go/logger/context_with_logging.go
@@ -6,53 +6,53 @@ import (
 	"golang.org/x/net/context"
 )
 
-type ContextAndLogger struct {
+type Context struct {
 	ctx context.Context
 	Logger
 }
 
-var _ CtxAndLogger = ContextAndLogger{}
+var _ CtxAndLogger = Context{}
 
-func (c ContextAndLogger) Ctx() context.Context {
+func (c Context) Ctx() context.Context {
 	return c.ctx
 }
 
-func NewContexAndLogger(c context.Context, l Logger) ContextAndLogger {
-	return ContextAndLogger{ctx: c, Logger: l}
+func NewContext(c context.Context, l Logger) Context {
+	return Context{ctx: c, Logger: l}
 }
 
-func (c ContextAndLogger) UpdateContext(ctx context.Context) CtxAndLogger {
-	return NewContexAndLogger(ctx, c.Logger)
+func (c Context) UpdateContext(ctx context.Context) CtxAndLogger {
+	return NewContext(ctx, c.Logger)
 }
 
-func (c ContextAndLogger) Debug(format string, args ...interface{}) {
+func (c Context) Debug(format string, args ...interface{}) {
 	c.Logger.CloneWithAddedDepth(1).CDebugf(c.ctx, format, args)
 }
 
-func (c ContextAndLogger) Info(format string, args ...interface{}) {
+func (c Context) Info(format string, args ...interface{}) {
 	c.Logger.CloneWithAddedDepth(1).CInfof(c.ctx, format, args)
 }
 
-func (c ContextAndLogger) Notice(format string, args ...interface{}) {
+func (c Context) Notice(format string, args ...interface{}) {
 	c.Logger.CloneWithAddedDepth(1).CNoticef(c.ctx, format, args)
 }
 
-func (c ContextAndLogger) Warning(format string, args ...interface{}) {
+func (c Context) Warning(format string, args ...interface{}) {
 	c.Logger.CloneWithAddedDepth(1).CWarningf(c.ctx, format, args)
 }
 
-func (c ContextAndLogger) Error(format string, args ...interface{}) {
+func (c Context) Error(format string, args ...interface{}) {
 	c.Logger.CloneWithAddedDepth(1).CErrorf(c.ctx, format, args)
 }
 
-func (c ContextAndLogger) Critical(format string, args ...interface{}) {
+func (c Context) Critical(format string, args ...interface{}) {
 	c.Logger.CloneWithAddedDepth(1).CCriticalf(c.ctx, format, args)
 }
 
-func (c ContextAndLogger) Fatal(format string, args ...interface{}) {
+func (c Context) Fatal(format string, args ...interface{}) {
 	c.Logger.CloneWithAddedDepth(1).CFatalf(c.ctx, format, args)
 }
 
 func NewContextTodoWithLoggingForTesting(t *testing.T) CtxAndLogger {
-	return &ContextAndLogger{ctx: context.TODO(), Logger: NewTestLogger(t)}
+	return &Context{ctx: context.TODO(), Logger: NewTestLogger(t)}
 }
